refactor: remove type cast; tweak handle filter
Instead of copying criteria-matching
SafeHandleEx instances to a list of
SafeFileHandleEx, we cast handles straight
to SafeFileHandleEx and remove those
which are do not match the criteria.

The criteria are unchanged.
1. A handle's object 's Type must be either
unidentified or "File" (file/directory).
2. If the object type is a File and we
succeed in querying the file path , the
path must contain the query string.
3. If we fail to query the Type or the file
path, we do not remove it from the
results.