refactor: update HandlesFilter enum with IncludeProtectedProcesses
! This may need to be update again to differentiate between different levels of process protection.

This new enum value is intended to include handles from processes with Light or Full process protection.
This protection limits the information that can be acquired from a process e.g. command line, per-handle details, and more.
Among these details is the Name of a handle (i.e. the full file path) and the Type of a handle's object, hence the inherent inclusion of IncludeFailedTypeQuery.